requirements.txt
numpy==1.24.2 
pandas==1.5.3
pathlib==1.0.1
opencv-python==4.7.0.68
pytesseract==0.3.10 #you will need to download tesseract seperately
scipy==1.10.0
Pillow==9.4.0
selenium==4.8.0
nltk==3.8.1
torch==1.13.1
sentence-transformers==2.2.2
pyspellchecker==0.7.1